### 2020-10-25 #2777 @vlandau
Timothee Poisot (tpoisot) has offered to review this paper/software. From the spreadsheet of potential reviewers, arbennett, r-barnes, and juliohm look like excellent fits.
