+++ README.md
@@ -19,8 +19,8 @@ using Pkg; Pkg.add(PackageSpec(name = "Omniscape", rev = "main"))
 
 ## Citing Omniscape.jl
 
-A formal paper detailing Omniscape.jl is forthcoming, but until it is published, please use the something like the following to cite Omniscape.jl if you use it in your work:
-> Landau, V. A. 2020. Omniscape.jl: An efficient and scalable implementation of the Omniscape algorithm in the Julia programming language, vX.Y.Z, URL: https://github.com/Circuitscape/Omniscape.jl, DOI: 10.5281/zenodo.3955123.
+A formal paper detailing Omniscape.jl is currently [in pre-review](https://github.com/openjournals/joss-reviews/issues/2777). Until it is published, please use the following to cite Omniscape.jl if you use it in your work:
+> Landau, V. A. 2020. Omniscape.jl: Software to compute omnidirectional landscape connectivity, vX.Y.Z, URL: https://github.com/Circuitscape/Omniscape.jl, DOI: 10.5281/zenodo.3955123.
 
 Here's a bibtex entry:
 ```
