### 2020-10-29 145b6d5
change all-to-one to advanced in docs
