### 2020-10-29 bb85c41
change all-to-one to advanced in paper
