### 2020-10-29 ef80ff1
remove type def that is not used (#82)
