### 2020-10-29 ab5217e
Vl/acknowl (#83)
add acknowledgments to readme and docs
