### 2020-10-29 d550042
fix bibtex entry for citation
