### 2020-11-02 #2777 @melissawm
:wave: @tpoisot, @juliohm would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html Let me know if you have any further questions. Thank you!
