+++ paper/paper.bib
@@ -2,6 +2,7 @@
   title={Isolation by resistance},
   author={McRae, Brad H},
   journal={Evolution},
+  doi={10.1111/j.0014-3820.2006.tb00500.x},
   volume={60},
   number={8},
   pages={1551--1561},
@@ -11,6 +12,7 @@
 
 @article{bezanson2017,
   title={Julia: A fresh approach to numerical computing},
+  doi={10.1137/141000671},
   author={Bezanson, Jeff and Edelman, Alan and Karpinski, Stefan and Shah, Viral B},
   journal={SIAM review},
   volume={59},
@@ -24,6 +26,7 @@
   title={Using circuit theory to model connectivity in ecology, evolution, and conservation},
   author={McRae, Brad H and Dickson, Brett G and Keitt, Timothy H and Shah, Viral B},
   journal={Ecology},
+  doi={10.1890/07-1861.1},
   volume={89},
   number={10},
   pages={2712--2724},
@@ -33,6 +36,7 @@
 
 @article{dickson2019,
   title={Circuit-theory applications to connectivity science and conservation},
+  doi={10.1890/07-1861.1},
   author={Dickson, Brett G and Albano, Christine M and Anantharaman, Ranjan and Beier, Paul and Fargione, Joe and Graves, Tabitha A and Gray, Miranda E and Hall, Kimberly R and Lawler, Josh J and Leonard, Paul B and others},
   journal={Conservation Biology},
   volume={33},
