+++ paper/paper.bib
@@ -36,7 +36,7 @@
 
 @article{dickson2019,
   title={Circuit-theory applications to connectivity science and conservation},
-  doi={10.1890/07-1861.1},
+  doi={10.1111/cobi.13230},
   author={Dickson, Brett G and Albano, Christine M and Anantharaman, Ranjan and Beier, Paul and Fargione, Joe and Graves, Tabitha A and Gray, Miranda E and Hall, Kimberly R and Lawler, Josh J and Leonard, Paul B and others},
   journal={Conservation Biology},
   volume={33},
