### 2020-11-06 #2777 @tpoisot
*grabs :microphone:  to give some completely unsolicited advice*

I'm happy to review this one, I used Omniscape a lot. I know that @vlucet and @glaroc also did, and would be suitable reviewers as well.
