### 2020-11-06 #2777 @VLucet
Thanks @tpoisot for thinking of me! I actually have not had the chance to test Omniscape yet. I know @glaroc did and would therefore be better suited. I have used Circuitscape (Julia version) quite a bit. I'm unsure whether having this perspective fully qualifies me to be a reviewer.
