### 2020-11-09 #2829 @whedon
[REVIEW]: Omniscape.jl: Software to predict omni-directional landscape connectivity
**Submitting author:** @vlandau (<a href="http://orcid.org/0000-0001-9290-9438">Vincent Landau</a>)
**Repository:** <a href="https://github.com/Circuitscape/Omniscape.jl" target ="_blank">https://github.com/Circuitscape/Omniscape.jl</a>
**Version:** v0.4.4
