### 2020-11-18 #2829 @melissawm
Yes, the paper should be published under the new title, but I might need to check this upon acceptance. Can you please remind me to do that, just in case? :) Thanks!
