### 2020-11-18 #2829 @juliohm
@vlandau can you please provide an example in the documentation that can be run easily? The example provided depends on  a file that is not available. Also, can you plot the output of this example in the documentation? Ideally this example should come as one of the first things in your docs before you even start explaining the algorithm.

Also, can you please add a statement of need explaining why this functionality is important, and how other packages from other languages may provide similar functionality?
