### 2020-11-18 #2829 @tpoisot
I concur with @juliohm  - it would be good to have a fully runnable example, and the results shown in the paper.
