### 2020-11-18 #2829 @vlandau
@juliohm yes definitely! I tried to get at the importance of the functionality of the package in the second paragraph in the motivation section -- I explain the value of the Omniscape algorithm, how it differs from the similar package "Circuitscape.jl", and that until now there was not a software package implementing it. Should I just shift that to its own "Statement of Need" section?
