### 2020-11-18 #2829 @juliohm
I think I'm missing the context here. Someone coming for the first time in
the subject will wonder why ominiscape ? What problems it addresses? You
need to show the potential users what it is about and how the features are
used in practical applications.
