### 2020-11-18 c1ec105
add working example to docs
