### 2020-11-19 ee6ab00
add example with plotted output -- first pass...
