### 2020-11-19 f69d6db
update usage section to link to examples
