### 2020-11-19 e94a424
some reworking of docs
