+++ docs/make.jl
@@ -16,6 +16,7 @@ makedocs(
     sitename = "Omniscape.jl",
     pages = ["Home" => "index.md",
     		 "User Guide" => "usage.md",
+			 "Function Documentation" => "apidocs.md",
 			 "Examples" => "examples.md"],
 )
 
