+++ docs/src/apidocs.md
@@ -1,0 +1,4 @@
+# Function Documentation
+```@docs
+run_omniscape
+```
