+++ latest/search_index.js
@@ -1,3 +1,3 @@
 var documenterSearchIndex = {"docs":
-[{"location":"usage/#User-Guide","page":"User Guide","title":"User Guide","text":"","category":"section"},{"location":"usage/#Installation","page":"User Guide","title":"Installation","text":"","category":"section"},{"location":"usage/","page":"User Guide","title":"User Guide","text":"The latest version of Omniscape.jl requires Julia version 1.5 or greater. You can install Julia here. Once installation is complete, open a Julia terminal and run the following code to install Omniscape.jl.","category":"page"},{"location":"usage/","page":"User Guide","title":"User Guide","text":"using Pkg; Pkg.add(\"Omniscape\")","category":"page"},{"location":"usage/","page":"User Guide","title":"User Guide","text":"If you want to install the latest (unreleased) development version of Omniscape.jl, you can get it by running:","category":"page"},{"location":"usage/","page":"User Guide","title":"User Guide","text":"using Pkg; Pkg.add(PackageSpec(name = \"Omniscape\", rev = \"main\"))","category":"page"},{"location":"usage/#Running-Omniscape","page":"User Guide","title":"Running Omniscape","text":"","category":"section"},{"location":"usage/","page":"User Guide","title":"User Guide","text":"For detailed examples of how to use Omniscape, check out the Examples section. Omniscape.jl provides a single user-facing function: run_omniscape.","category":"page"},{"location":"usage/","page":"User Guide","title":"User Guide","text":"run_omniscape() offers two methods. The first, shown above, accepts the path to an INI file that specifies file paths for raster inputs and other user-specified options. An INI file can be created using any text editor (e.g. notepad) and saved with the .ini file extension. The following code block shows an example INI file. The headings in square brackets are not required. They are there for organization purposes and are ignored by run_omniscape().","category":"page"},{"location":"usage/","page":"User Guide","title":"User Guide","text":"[Required arguments]\nresistance_file = resistance_surface.tif\nradius = 100\nblock_size = 5\nproject_name = output/example\n\n[General options]\nsource_from_resistance = true\nr_cutoff = 50\ncalc_normalized_current = true\n\nparallelize = true\nparallel_batch_size = 20\n\n[Output options]\nwrite_raw_currmap = true","category":"page"},{"location":"usage/","page":"User Guide","title":"User Guide","text":"The second method of run_omniscape accepts in-memory objects representing resistance and other spatial data inputs, and a dictionary of arguments specifying Omniscape options.","category":"page"},{"location":"usage/","page":"User Guide","title":"User Guide","text":"The full suite of arguments that can be specified in the .ini file are described in detail in Arguments.","category":"page"},{"location":"usage/#Parallel-Processing","page":"User Guide","title":"Parallel Processing","text":"","category":"section"},{"location":"usage/","page":"User Guide","title":"User Guide","text":"Omniscape uses parallel processing by default, but currently, Julia requires that the number of parallel threads to use be specified by an environment variable called JULIA_NUM_THREADS. This environment variable needs to be defined prior to launching Julia. The following examples demonstrate how to set JULIA_NUM_THREADS and start up Julia to use 4 threads from your terminal.","category":"page"},{"location":"usage/","page":"User Guide","title":"User Guide","text":"On Linux/Mac","category":"page"},{"location":"usage/","page":"User Guide","title":"User Guide","text":"export JULIA_NUM_THREADS=4\njulia","category":"page"},{"location":"usage/","page":"User Guide","title":"User Guide","text":"On Windows","category":"page"},{"location":"usage/","page":"User Guide","title":"User Guide","text":"set JULIA_NUM_THREADS=4\njulia","category":"page"},{"location":"usage/#Arguments","page"
[…]
