### 2020-11-19 fbb6fd4
rearrange docs, fix map figure in examples
