### 2020-11-19 #2829 @vlandau
@juliohm @tpoisot please see the new [Examples section](https://docs.circuitscape.org/Omniscape.jl/latest/examples/#Examples) in the docs. Following the example set by other Julia packages' documentation, I created a dedicated section for examples and linked to it in both the table of contents and the User Guide.
