### 2020-11-20 #2829 @juliohm
@vlandau the example section you linked could be improved. Imagine that you are explaining this subject to someone that never heard of connectivity maps before. You need to plot the input images, display the entries of the resistance table, and how they can be interpreted, and finally display the result of the `run_omniscape` function, ideally without using a file, but instead the method with arguments. The easier you make the life of the reader, the more they will appreciate your work. Can you refactor the example like that?
