### 2020-11-20 #2829 @vlandau
@juliohm thanks for the feedback -- I can add those details, and good idea to use the alternative method instead of the INI file -- then the only thing that needs to be downloaded is the land cover data. I'll post back here when I have made a second pass.
