### 2020-11-24 3a7c2d4
Update examples.md
