### 2020-11-24 fa2d157
Update examples.md
