+++ latest/search_index.js
@@ -1,3 +1,3 @@
 var documenterSearchIndex = {"docs":
-[{"location":"apidocs/#Function-Documentation","page":"Function Documentation","title":"Function Documentation","text":"","category":"section"},{"location":"apidocs/","page":"Function Documentation","title":"Function Documentation","text":"run_omniscape","category":"page"},{"location":"apidocs/#Omniscape.run_omniscape","page":"Function Documentation","title":"Omniscape.run_omniscape","text":"INI method:\n\nrun_omniscape(path::String)\n\nIn-memory method:\n\nrun_omniscape(\n    cfg::Dict{String, String}\n    resistance::Array{Union{Float64, Missing}, 2};\n    reclass_table = Array{Union{Float64, Missing}, 2}(undef, 1, 2),\n    source_strength = source_from_resistance(resistance, cfg, reclass_table),\n    condition1 = Array{Union{Float64, Missing}, 2}(undef, 1, 1),\n    condition2 = Array{Union{Float64, Missing}, 2}(undef, 1, 1),\n    condition1_future = Array{Union{Float64, Missing}, 2}(undef, 1, 1),\n    condition2_future = Array{Union{Float64, Missing}, 2}(undef, 1, 1),\n    wkt = \"\",\n    geotransform = [0.0, 1.0, 0.0, 0.0, 0.0, -1.0],\n    write_outputs = false\n)\n\nCompute omnidirectional current flow. All array inputs for the in-memory method should be of type Array{Union{T, Missing}, 2} where T <: Number, with missing used for NoData pixels.\n\nParameters\n\npath: The path to an INI file containing run parameters. See the Arguments section of the User Guide for descriptions of the run paramters.\n\ncfg: A dictionary of Omniscape run parameters. See the Arguments section of the User Guide for descriptions of the run parameters and their default values. The in-memory method of run_omniscape ignores the following keys: resistance_file, source_file, reclass_table, condition1_file, condition2_file, condition1_future_file, and condition2_future_file. These all specify file paths, so they do not apply to the in-memory method of run_omniscape.\n\nresistance: An 2D, north-oriented array of resistance values. Use missing for NoData (infinite resistance). resistance cannot contain zeros or negative values.\n\nKeyword Arguments\n\nreclass_table:  A two column array. The first column contains the original resistance values in the resistance surface, and the second column specifies what those values should be changed to. You can reclassify values to missing to replace them with infinite resistance (NoData).\n\nsource_strength: A 2D, north-oriented array (with size equal to size(resistance)) of source strength values. source_strength is only required if source_from_resistance in cfg is set to \"false\" (the default value).\n\ncondition1: Optional. Required if conditional incfg is set to \"true\". A 2D, north-oriented array (with size equal to size(resistance)). See Climate Connectivity and Conditional Connectivity Options for more information.\n\ncondition2: Optional. Required if conditional incfg is set to \"true\" and n_conditions in cfg is set to \"2\". A 2D, north-oriented array (with size equal to size(resistance)). See Climate Connectivity and Conditional Connectivity Options for more information.\n\ncondition1_future: Optional. Required if conditional in cfg is set to \"true\" and compare_to_future in cfg is set to \"1\" or \"both\". A 2D, north-oriented array (with size equal to size(resistance)). See Climate Connectivity and Conditional Connectivity Options for more information.\n\ncondition2_future: Optional. A 2D, north-oriented array (with size equal to size(resistance)). See Climate Connectivity and Conditional Connectivity Options for more information. Required if conditional incfg is set to \"true\", n_conditions in cfg is set to \"2\", and compare_to_future in cfg is set to \"2\" or \"both\".\n\nwkt: Optionally specify a Well Known Text representation of the projection use
[…]
