### 2020-11-30 #2829 @vlandau
Hi @juliohm and @tpoisot -- I've updated the example in the documentation to be much more thorough. I have not included it in the paper yet since I'm at (or over) the soft 1000 word limit for JOSS papers and the docs are linked in the paper, but happy to add it if you two prefer that. [Here](https://docs.circuitscape.org/Omniscape.jl/latest/examples/#Forest-connectivity-in-central-Maryland) is the expanded example. Thanks!
