### 2020-11-30 #2829 @juliohm
I still miss the input image and the intermediate results of the example.
Are you aware that Documenter.jl allows you to run code directly? You don't
need to include code snippets and copy paste results. Just add the actual
script section by section and the results will be shown. Check the
GeoStats.jl docs for examples. In particular the Basic Worfklow section.
