### 2020-11-30 #2829 @vlandau
@juliohm Yeah I was aware of that but had been trying to avoid that since the example takes a while to run on Travis workers. I totally missed the request to see a plot of the input landcover data! Sorry about that. I'm not sure what you mean about intermediate results, though?
