### 2020-11-30 #2829 @juliohm
By intermediate results I mean the input image and any intermediate output
produced. If the example takes a long time to run you probably need to
reduce the image? This is supposed to be a small toy example to give users
an idea of what is happening. It doesn't need to be an actual real world
application with a large image. Makes sense?
