### 2020-11-30 #2829 @vlandau
Okay -- I'll reduce the image size, and will have Documenter.jl run the code directly and have it show all of the REPL outputs. I'll let you know when it's done. Thanks!
