### 2020-11-30 #2829 @vlandau
@juliohm do you prefer `@repl` blocks or `@example` blocks? I see you use the former in GeoStats.jl docs.
