### 2020-11-30 #2829 @juliohm
I think @example blocks are fine. As soon as the message is clear it
doesn't matter really.
