### 2020-11-30 2c23098
run the Julia code for examples in docs generation
