+++ docs/src/examples.md
@@ -6,29 +6,40 @@ Land cover datasets are commonly used to parameterize resistance for connectivit
 
 First, install the necessary packages and import them:
 
-```julia
+```@example mdforest
 using Pkg; Pkg.add(["Omniscape", "GeoData", "Plots"])
 using Omniscape, GeoData, Plots
 ```
 
-Next, download the landcover data we'll use in this example:
+Next, download the landcover data we'll use in this example, and plot it:
 
-```julia
+```@example mdforest
 url_base = "https://raw.githubusercontent.com/Circuitscape/datasets/main/"
 # Download the NLCD tile used to create the resistance surface and load it
 download(string(url_base, "data/nlcd_2016_frederick_md.tif"),
          "nlcd_2016_frederick_md.tif")
 
-# Load the array using one of Omniscape's internal functions, or a function
-# from a GIS Julia package of your choice. Omniscape's `read_raster()` returns a
-# tuple with the data array, a wkt string containing geographic projection info,
-# and an array containing geotransform values.
+# Plot the landcover data
+values = [11, 21, 22, 23, 24, 31, 41, 42, 43, 52, 71, 81, 82, 90, 95]
+palette = ["#476BA0", "#DDC9C9", "#D89382", "#ED0000", "#AA0000",
+           "#b2b2b2", "#68AA63", "#1C6330", "#B5C98E", "#CCBA7C",
+           "#E2E2C1", "#DBD83D", "#AA7028", "#BAD8EA", "#70A3BA"]
+
+plot(GeoArray(GDALarray("nlcd_2016_frederick_md.tif")),
+     title = "Land Cover Type", xlabel = "Easting", ylabel = "Northing",
+     seriescolor = cgrad(palette, (values .- 12) ./ 84, categorical = true),
+     size = (600, 550))
+```
+
+Now, load the array using Omniscape's internal `read_raster()` function or a function from a GIS Julia package of your choice. `read_raster()` returns a tuple with the data array, a wkt string containing geographic projection info, and an array containing geotransform values. We'll use the wkt and geotransform later.
+
+```@example mdforest
 land_cover, wkt, transform = Omniscape.read_raster("nlcd_2016_frederick_md.tif", Float64)
 ```
 
 The next step is to create a resistance reclassification table that defines a resistance value for each land cover value. Land cover values go in the left column, and resistance values go in the right column. In this case, we are modeling forest connectivity, so forest classes receive the lowest resistance score of one. Other "natural" land cover types are assigned moderate values, and human-developed land cover types are assigned higher values. Medium- to high-intensity development are given a value of `missing`, which denotes infinite resistance (absolute barriers to movement).
 
-```julia
+```@example mdforest
 # Create the reclassification table used to translate land cover into resistance
 reclass_table = [
     11.	100; # Water
@@ -51,7 +62,7 @@ reclass_table = [
 
 Next, we define the configuration options for this model run. See the [Arguments](@ref) section in the [User Guide](@ref) for more information about each of the configuration options.
 
-```julia
+```@example mdforest
 # Specify the configuration options
 config = Dict{String, String}(
     "radius" => "100",
@@ -63,9 +74,9 @@ config = Dict{String, String}(
 )
 ```
 
-Finally, run Omniscape, feeding in the configuration dictionary, the resistance array, the reclass table, as well as the wkt and geotransform information loaded above with `Omniscape.read_raster()`. Passing in the wkt and geotransform, along with `true` for the `write_outputs` argument, will allow Omniscape to write the outputs as properly projected rasters.
[…]
