### 2020-11-30 df95816
address deprecated edit_branch, and add ENV def to fix GR warnings
