+++ latest/md_nlcd_omniscape_output/config.ini
@@ -1,0 +1,6 @@
+radius = 100
+source_from_resistance = true
+block_size = 21
+r_cutoff = 1
+project_name = md_nlcd_omniscape_output
+reclassify_resistance = true
