### 2020-11-30 0a44da2
clean up some stuff in example, increase plot sizes
