+++ docs/src/examples.md
@@ -6,9 +6,13 @@ Land cover datasets are commonly used to parameterize resistance for connectivit
 
 First, install the necessary packages and import them:
 
-```@example mdforest
+```julia
 using Pkg; Pkg.add(["Omniscape", "GeoData", "Plots"])
 using Omniscape, GeoData, Plots
+```
+```julia
+using Pkg; Pkg.add(["Omniscape", "GeoData", "Plots"]) # hide
+using Omniscape, GeoData, Plots # hide
 nothing # hide
 ```
 
@@ -28,8 +32,7 @@ palette = ["#476BA0", "#DDC9C9", "#D89382", "#ED0000", "#AA0000",
 
 plot(GeoArray(GDALarray("nlcd_2016_frederick_md.tif")),
      title = "Land Cover Type", xlabel = "Easting", ylabel = "Northing",
-     seriescolor = cgrad(palette, (values .- 12) ./ 84, categorical = true),
-     size = (2400, 2200))
+     seriescolor = cgrad(palette, (values .- 12) ./ 84, categorical = true))
 ```
 
 Now, load the array using Omniscape's internal `read_raster()` function or a function from a GIS Julia package of your choice. `read_raster()` returns a tuple with the data array, a wkt string containing geographic projection info, and an array containing geotransform values. We'll use the wkt and geotransform later.
@@ -94,6 +97,5 @@ Now, load the current map back into Julia as spatial data and plot it:
 current = GDALarray("md_nlcd_omniscape_output/cum_currmap.tif")
 plot(current,
      title = "Cumulative Current Flow", xlabel = "Easting", ylabel = "Northing",
-     seriescolor = cgrad(:inferno, [0, 0.005, 0.03, 0.06, 0.1, 0.15]),
-     size = (2400, 2200))
+     seriescolor = cgrad(:inferno, [0, 0.005, 0.03, 0.06, 0.1, 0.15]))
 ```
