### 2020-11-30 d58f459
add figure caption to current map
