+++ docs/src/examples.md
@@ -99,4 +99,4 @@ plot(current,
      title = "Cumulative Current Flow", xlabel = "Easting", ylabel = "Northing",
      seriescolor = cgrad(:inferno, [0, 0.005, 0.03, 0.06, 0.1, 0.15]))
 ```
-*Cumulative current flow representing foret connectivity. Note that areas in white correspond to build up areas (NLCD values of 23 and 24) that act as absolute barriers to movement.*
+*Cumulative current flow representing foret connectivity. Note that areas in white correspond to built up areas (NLCD values of 23 and 24) that act as absolute barriers to movement.*
