### 2020-11-30 65b5cfe
update plots to use saved figures
