+++ docs/src/examples.md
@@ -32,8 +32,12 @@ palette = ["#476BA0", "#DDC9C9", "#D89382", "#ED0000", "#AA0000",
 
 plot(GeoArray(GDALarray("nlcd_2016_frederick_md.tif")),
      title = "Land Cover Type", xlabel = "Easting", ylabel = "Northing",
-     seriescolor = cgrad(palette, (values .- 12) ./ 84, categorical = true))
+     seriescolor = cgrad(palette, (values .- 12) ./ 84, categorical = true),
+     size = (600, 550))
+savefig("nlcd-plot.svg"); nothing # hide
 ```
+
+<img src='nlcd-plot.svg' width=500><br>
 
 Now, load the array using Omniscape's internal `read_raster()` function or a function from a GIS Julia package of your choice. `read_raster()` returns a tuple with the data array, a wkt string containing geographic projection info, and an array containing geotransform values. We'll use the wkt and geotransform later.
 
@@ -97,6 +101,10 @@ Now, load the current map back into Julia as spatial data and plot it:
 current = GDALarray("md_nlcd_omniscape_output/cum_currmap.tif")
 plot(current,
      title = "Cumulative Current Flow", xlabel = "Easting", ylabel = "Northing",
-     seriescolor = cgrad(:inferno, [0, 0.005, 0.03, 0.06, 0.1, 0.15]))
+     seriescolor = cgrad(:inferno, [0, 0.005, 0.03, 0.06, 0.1, 0.15]),
+     size = (600, 550))
+savefig("current-plot.svg"); nothing # hide
 ```
-*Cumulative current flow representing foret connectivity. Note that areas in white correspond to built up areas (NLCD values of 23 and 24) that act as absolute barriers to movement.*
+
+<img src='current-plot.svg' width=500> <br><em>Cumulative current flow representing forest connectivity. Note that areas in white correspond to built up areas (NLCD values of 23 and 24) that act as absolute barriers to movement.</em><br><br>
+
