+++ docs/src/examples.md
@@ -34,10 +34,10 @@ plot(GeoArray(GDALarray("nlcd_2016_frederick_md.tif")),
      title = "Land Cover Type", xlabel = "Easting", ylabel = "Northing",
      seriescolor = cgrad(palette, (values .- 12) ./ 84, categorical = true),
      size = (600, 550))
-savefig("nlcd-plot.svg"); nothing # hide
+savefig("nlcd-plot.png"); nothing # hide
 ```
 
-<img src='nlcd-plot.svg' width=500><br>
+<img src='nlcd-plot.png' width=500><br>
 
 Now, load the array using Omniscape's internal `read_raster()` function or a function from a GIS Julia package of your choice. `read_raster()` returns a tuple with the data array, a wkt string containing geographic projection info, and an array containing geotransform values. We'll use the wkt and geotransform later.
 
@@ -103,8 +103,8 @@ plot(current,
      title = "Cumulative Current Flow", xlabel = "Easting", ylabel = "Northing",
      seriescolor = cgrad(:inferno, [0, 0.005, 0.03, 0.06, 0.1, 0.15]),
      size = (600, 550))
-savefig("current-plot.svg"); nothing # hide
+savefig("current-plot.png"); nothing # hide
 ```
 
-<img src='current-plot.svg' width=500> <br><em>Cumulative current flow representing forest connectivity. Note that areas in white correspond to built up areas (NLCD values of 23 and 24) that act as absolute barriers to movement.</em><br><br>
+<img src='current-plot.png' width=500> <br><em>Cumulative current flow representing forest connectivity. Note that areas in white correspond to built up areas (NLCD values of 23 and 24) that act as absolute barriers to movement.</em><br><br>
 
