### 2020-12-01 494f820
have docs use saved figs instead as workaround
