+++ docs/src/examples.md
@@ -13,12 +13,15 @@ using Omniscape, GeoData, Plots
 ```@setup mdforest
 using Pkg; Pkg.add(["Omniscape", "GeoData", "Plots"])
 using Omniscape, GeoData, Plots
+url_base = "https://raw.githubusercontent.com/Circuitscape/datasets/main/"
+download(string(url_base, "data/nlcd_2016_frederick_md.tif"),
+         "nlcd_2016_frederick_md.tif")
 nothing
 ```
 
 Next, download the landcover data we'll use in this example, and plot it:
 
-```@example mdforest
+```julia
 url_base = "https://raw.githubusercontent.com/Circuitscape/datasets/main/"
 # Download the NLCD tile used to create the resistance surface and load it
 download(string(url_base, "data/nlcd_2016_frederick_md.tif"),
@@ -33,11 +36,10 @@ palette = ["#476BA0", "#DDC9C9", "#D89382", "#ED0000", "#AA0000",
 plot(GeoArray(GDALarray("nlcd_2016_frederick_md.tif")),
      title = "Land Cover Type", xlabel = "Easting", ylabel = "Northing",
      seriescolor = cgrad(palette, (values .- 12) ./ 84, categorical = true),
-     size = (600, 550))
-savefig("nlcd-plot.png"); nothing # hide
+     size = (700, 640))
 ```
 
-<img src='nlcd-plot.png' width=500><br>
+<img src='../figs/mdlc.png' width=500><br>
 
 Now, load the array using Omniscape's internal `read_raster()` function or a function from a GIS Julia package of your choice. `read_raster()` returns a tuple with the data array, a wkt string containing geographic projection info, and an array containing geotransform values. We'll use the wkt and geotransform later.
 
@@ -97,14 +99,13 @@ You'll see that outputs are written to a new folder called "md\_nlcd\_omniscape\
 
 Now, load the current map back into Julia as spatial data and plot it:
 
-```@example mdforest
+```julia
 current = GDALarray("md_nlcd_omniscape_output/cum_currmap.tif")
 plot(current,
      title = "Cumulative Current Flow", xlabel = "Easting", ylabel = "Northing",
      seriescolor = cgrad(:inferno, [0, 0.005, 0.03, 0.06, 0.1, 0.15]),
      size = (600, 550))
-savefig("current-plot.png"); nothing # hide
 ```
 
-<img src='current-plot.png' width=500> <br><em>Cumulative current flow representing forest connectivity. Note that areas in white correspond to built up areas (NLCD values of 23 and 24) that act as absolute barriers to movement.</em><br><br>
+<img src='../figs/md-curmap.png' width=500> <br><em>Cumulative current flow representing forest connectivity. Note that areas in white correspond to built up areas (NLCD values of 23 and 24) that act as absolute barriers to movement.</em><br><br>
 
