### 2020-12-01 e6d9551
adjust color ramp in curmap plot
