### 2020-12-01 aa9088e
first attempt at GitHub Actions CI
