+++ .travis.yml
@@ -1,34 +1,0 @@
-language: julia
-
-env:
-  - JULIA_NUM_THREADS=2
-
-os:
-  - linux
-
-julia:
-  - 1.5
-
-notifications:
-  email: false
-
-jobs:
-  include:
-    - stage: "Documentation"
-      os: linux
-      julia: 1.5
-      script:
-        - julia --project=docs/ -e 'using Pkg; Pkg.develop(PackageSpec(path=pwd()));
-                                               Pkg.instantiate()'
-        - travis_wait 30 julia --project=docs/ docs/make.jl
-    - stage: "Code coverage"
-      if: branch = main
-      julia: 1.5
-      os: linux
-      script:
-        - julia -e 'using Pkg;
-          Pkg.build(); Pkg.test(coverage=true);
-          import Omniscape; cd(joinpath(dirname(pathof(Omniscape)),".."));
-          Pkg.add("Coverage");
-          using Coverage;
-          Codecov.submit(Codecov.process_folder())'
