+++ .github/workflows/ci.yml
@@ -18,7 +18,6 @@ jobs:
       matrix:
         version:
           - '1.5'
-          - '1'
         os:
           - ubuntu-latest
           - macos-latest
