### 2020-12-01 #2829 @vlandau
@juliohm @tpoisot [here]() is the latest iteration of the example. @juliohm, I ended up needing to add the plots as figures (saved in the docs directory before hand) because I was running into various issues when trying to plot them dynamically (figures being too small, Documenter hanging at the plot steps, etc). Hopefully this is getting closer! :slightly_smiling_face:
