### 2020-12-08 #2829 @juliohm
Regarding the Documenter hanging at the plot steps, perhaps you are experiencing the same issue as I am with the latest stable release of Plots.jl? https://github.com/JuliaPlots/Plots.jl/issues/3164 Comment there if you can reproduce the issue. Otherwise just ignore this comment.
