### 2020-12-08 #2829 @juliohm
![image](https://user-images.githubusercontent.com/3345261/101531088-888d9580-3971-11eb-8160-c80b620ce41e.png)

You claim 3 different outputs in the section above. Can you please show this functionality in the example as well? Would it be easy to produce the two other optional outputs?
