### 2020-12-08 #2829 @juliohm
Before I continue with the review of the paper, please consider addressing the two issues raised in previous comments about the package documentation:

1. A clear statement of need must be present in the very beginning of the documentation. Your documentation is currently starting with the description of an algorithm without any motivation or description of the problem being solved. Add a simple sentence like "In environmental sciences, ecology, etc, it is often important to discover or quantify connectivity of forests. This task has various challenges and this package is a possible solution to this." Then, after you state the problem you are addressing, you can jump to the example. I don't think that describing the algorithm is the first thing an end user wants. Keep in mind that the audience of a documentation is mostly interested in knowing if the package has a set of features that solves their problem. Only a few people will devote time to reading the algorithm explanation and an even smaller fraction of people will try to read the papers you cited.

2. Make sure that the example covers the full functionality of the package. Your example is must improved compared to previous versions but it still doesn't show the optional outputs described in the docs. I'd try to extend the example to show these outputs and comment on how they can be used in practice.
