### 2020-12-08 #2829 @juliohm
That is fine. Make sure that the links are emphasized and that the
algorithm explanation is a bonus for the curious reader. This will improve
the user/reader experience.

Em ter., 8 de dez. de 2020 às 16:48, Vincent Landau <
notifications@github.com> escreveu:

> Would it be easy to produce the two other optional outputs?
>
> Yes definitely! I can work on that this week. I'll also add a statement of
> need to the landing page of the docs. Additional examples may be included
> in the future, so if it is okay I'll leave the examples in a dedicated
> section and provide a link to that section immediately following the
> statement of need.
>
