### 2020-12-14 cfaf3d3
use cholmod for one of the tests
