+++ test/input/config3.ini
@@ -15,5 +15,6 @@ calc_normalized_current = false
 mask_nodata = true
 
 write_as_tif = false
+solver = cholmod
 
 parallelize = true
