### 2020-12-14 0865665
reworked and added some new warnings, check for unsupported arguments
