+++ test/input/config3.ini
@@ -18,3 +18,7 @@ write_as_tif = false
 solver = cholmod
 
 parallelize = true
+
+[Unsupported arguments to test that the warning is thrown]
+foo = something
+bar = or_other
