### 2020-12-14 ee0aa8a
update examples to show all outputs
