### 2020-12-14 930c6cd
fix Kim's middle initial in citation
