+++ latest/md_nlcd_omniscape_output/config.ini
@@ -1,3 +1,5 @@
+calc_normalized_current = true
+calc_flow_potential = true
 radius = 100
 source_from_resistance = true
 block_size = 21
