### 2020-12-14 fc123b4
fix figure for normalize current
