### 2020-12-14 caf4a7f
fix file path for moving window figure
