+++ docs/src/index.md
@@ -15,7 +15,7 @@ Depth = 2
 
 ## About Omniscape.jl
 
-Omniscape.jl implements the Omniscape connectivity algorithm developed by [McRae et al. (2016)](https://www.researchgate.net/publication/304842896_Conserving_Nature's_Stage_Mapping_Omnidirectional_Connectivity_for_Resilient_Terrestrial_Landscapes_in_the_Pacific_Northwest). This software package can be used to produce maps of omni-directional habitat connectivity useful for scientific research as well as landscape management and conservation. Omniscape.jl is built on [Circuitscape.jl](https://github.com/Circuitscape/Circuitscape.jl). It offers a unique approach to connectivity modeling, particularly among circuit theoretic methods, by allowing the sources, destinations, and intensity of animal movement or ecological flow (modeled as electrical current) to be informed by continuous spatial data (such as a habitat suitability map). This information is combined with other spatial information on landscape resistance to movement or flow to produce models of habitat connectivity. To learn about how the algorithm works, see [The Omniscape Algorithm](@ref).
+Omniscape.jl implements the Omniscape connectivity algorithm developed by [McRae et al. (2016)](https://www.researchgate.net/publication/304842896_Conserving_Nature's_Stage_Mapping_Omnidirectional_Connectivity_for_Resilient_Terrestrial_Landscapes_in_the_Pacific_Northwest). This software package can be used to produce maps of omni-directional habitat connectivity useful for scientific research as well as landscape management and conservation. Omniscape.jl is built on [Circuitscape.jl](https://github.com/Circuitscape/Circuitscape.jl). It offers a unique approach to connectivity modeling, particularly among circuit theoretic methods, by allowing the sources, destinations, and intensity of animal movement or ecological flow (modeled as electrical current) to be informed by continuous spatial data (such as a habitat suitability map). This information is combined with other spatial information on landscape resistance to movement or flow to produce models of habitat connectivity. To learn about how the algorithm works, see [The Omniscape Algorithm](@ref). Check out the [Examples](@ref) section for a step-by-step demonstration of how to use Omniscape.jl.
 
 ### Outputs
 
