### 2020-12-14 #2829 @vlandau
@juliohm @tpoisot finally got to updating the docs to reflect the comments above!
