### 2020-12-14 #2829 @juliohm
Very nice! The documentation is now targeting the right audience! Like how you explained succinctly the work in the first paragraph of the docs, still providing reference and links to other sections for more details. That is it 💯

Continuing with the review...
