### 2020-12-14 #2829 @juliohm
Review comments for paper:

![image](https://user-images.githubusercontent.com/3345261/102064459-146e4a00-3dd6-11eb-9436-444044825ac2.png)

![image](https://user-images.githubusercontent.com/3345261/102064595-42ec2500-3dd6-11eb-8817-55dc474656ed.png)

The comma in `"cores, "` should probably go outside the double quotes?

Overall, I am missing references to alternative approaches to the connectivity map problem, and missing how the circuit-theoretic approach fits into a broader category of methods. If you could maybe cite different packages from other languages doing something similar, that would already help. Are there Python, R, ... packages for connectivity maps?
