### 2020-12-14 #2829 @vlandau
@juliohm [commas go inside quotes](https://style.mla.org/punctuation-and-quotation-marks/). There are no "standard" software packages that I know of. I think people likely just use their favorite graph packages or GIS plugins to compute least cost paths or corridors. I'll add one or two additional references and a sentence about least cost paths so the reader at least knows where to look!
