### 2020-12-14 50fbdef
Fix header in how it works section
makes for correct table of contents
