### 2020-12-16 c9a188a
some paper edits
