+++ paper/paper.bib
@@ -10,6 +10,17 @@
   publisher={Wiley Online Library}
 }
 
+@article{bunn2000,
+  title={Landscape connectivity: a conservation application of graph theory},
+  author={Bunn, Andrew G and Urban, Dean L and Keitt, Timothy H},
+  journal={Journal of environmental management},
+  volume={59},
+  number={4},
+  pages={265--278},
+  year={2000},
+  publisher={Elsevier}
+}
+
 @article{bezanson2017,
   title={Julia: A fresh approach to numerical computing},
   doi={10.1137/141000671},
