### 2020-12-16 #2829 @vlandau
@juliohm I've added a mention of the other predominant modeling approach (least-cost paths) as well as a citation.
