### 2020-12-16 83952df
fix names in citations
