+++ paper/paper.bib
@@ -59,7 +59,7 @@
 
 @inproceedings{shah2008,
   title={Circuitscape: a tool for landscape ecology},
-  author={Shah, Viral B and McRae, BH},
+  author={Shah, Viral B and McRae, Brad H},
   booktitle={Proceedings of the 7th Python in Science Conference},
   volume={7},
   pages={62--66},
@@ -82,7 +82,7 @@
 
 @article{mcrae2016,
   title={Conserving nature’s stage: mapping omnidirectional connectivity for resilient terrestrial landscapes in the pacific northwest},
-  author={McRae, BH and Popper, K and Jones, A and Schindel, M and Buttrick, S and Hall, K and Unnasch, RS and Platt, J},
+  author={McRae, Brad H and Popper, K and Jones, A and Schindel, M and Buttrick, S and Hall, K R and Unnasch, R S and Platt, J},
   journal={The Nature Conservancy, Portland, Oregon},
   year={2016}
 }
