### 2020-12-21 #2829 @tpoisot
I can't seem to edit the comment at the top - that being said, I have used Omniscape extensively, and there is no doubt that it is production-ready, efficient, and user-friendly. If anything, the current paper is selling the software short a little bit. Omniscape can be deployed on HPC to really benefit from the built-in multi-threading, and the variety of input/output formats is really important to mention.

I would also show one map with a full Omniscape result, to see a quick illustration of what the different components of the output are.
