### 2020-12-21 #2829 @melissawm
@tpoisot let's try this once more. If even after accepting the invitation you can't edit, we might be looking at a bug in whedon or some other issue we might need to investigate.
