### 2021-01-01 a84e43a
organized tests and closes #86
