+++ src/main.jl
@@ -138,13 +138,14 @@ function run_omniscape(
     project_name = cfg["project_name"]
     file_format = os_flags.write_as_tif ? "tif" : "asc"
 
-    ## Set number of BLAS threads to 1
-    BLAS.set_num_threads(1)
+    ## Set number of BLAS threads to 1 when parallel processing
+    if os_flags.parallelize && nthreads() != 1
+        BLAS.set_num_threads(1)
+    end
 
     check_resistance_values(resistance) && return
 
     # Reclassify resistance layer
-    # TODO create a function in utils.jl, reclassify_resistance()
     if os_flags.reclassify
         reclassify_resistance!(resistance, reclass_table)
     end
