+++ .github/workflows/ci.yml
@@ -18,6 +18,7 @@ jobs:
       matrix:
         version:
           - '1.5'
+          - '1.6'
         os:
           - ubuntu-latest
           - macos-latest
