### 2021-01-11 a94b038
fix 1.6 version in ci.yml
