+++ .github/workflows/ci.yml
@@ -18,7 +18,7 @@ jobs:
       matrix:
         version:
           - '1.5'
-          - '1.6'
+          - '^1.6.0-0' 
         os:
           - ubuntu-latest
           - macos-latest
