### 2021-01-15 #2829 @tpoisot
@melissawm / @vlandau - yep, all works - my comments on the (really good) paper are here:

> If anything, the current paper is selling the software short a little bit. Omniscape can be deployed on HPC to really benefit from the built-in multi-threading, and the variety of input/output formats is really important to mention.
>
> I would also show one map with a full Omniscape result, to see a quick illustration of what the different components of the output are.
