### 2021-01-15 2ec980a
add figure of outputs demo and caption
