### 2021-01-15 3146a30
update figure 2 caption
