### 2021-01-15 dbeddf4
fix figure 2 caption
