+++ paper/paper.md
@@ -60,7 +60,7 @@ Steps 1-4 are repeated for every potential target pixel. The resulting current m
 
 In addition to cumulative current, Omniscape.jl also optionally provides two additional outputs: flow potential, and normalized cumulative current (\autoref{fig:outputs}). Flow potential represents current flow under "null" resistance conditions and demonstrates what current flow would look like when unconstrained by resistance and barriers. Flow potential is calculated exactly as cumulative current flow, but with resistance set to one for the entire landscape. Normalized cumulative current flow is calculated by dividing cumulative current flow by flow potential. Normalized current helps identify areas where current is impeded or channelized (e.g., more or less current than expected under null resistance conditions). High values mean current flow is channelized, and low values mean current is impeded.
 
-![An example of the three different Omniscape.jl outputs. Outputs are from the "Maryland Forest Connectivity" example in [the documentation](docs.circuitscape.org/Omniscape.jl/latest). Cumulative current flow shows the total current for each landscape pixel. Flow potential shows predicted flow under resistance-free conditions. Normalized current shows the degree to which a pixel has more or less current than expected under resistance-free conditions (cumulative current flow divided by flow potential). Each layer is visualized using a quantile stretch with 30 breaks. Axes show easting and northing coordinates for reference.\label{outputs}](fig2.png)
+![An example of the three different Omniscape.jl outputs. Outputs are from the "Maryland Forest Connectivity" example in the software documentation.. Cumulative current flow shows the total current for each landscape pixel. Flow potential shows predicted flow under resistance-free conditions. Normalized current shows the degree to which a pixel has more or less current than expected under resistance-free conditions (cumulative current flow divided by flow potential). Each layer is visualized using a quantile stretch with 30 breaks. Axes show easting and northing coordinates for reference.\label{outputs}](fig2.png)
 
 # Usage
 
