### 2021-01-15 3a186a8
add sentence on HPC/multithreading
