+++ paper/paper.md
@@ -35,7 +35,7 @@ bibliography: paper.bib
 
 # Summary
 
-Omniscape.jl is a software package that implements the Omniscape algorithm [@mcrae2016] to compute landscape connectivity. It is written in the Julia programming language [@bezanson2017] to be fast, scalable, and easy-to-use. Circuitscape.jl [@anantharaman2020], the package on which Omniscape.jl builds and expands, abstracts landscapes as two-dimensional electrical networks and solves for current flow. The current flow that results represents landscape connectivity. Omniscape.jl is novel in that it produces maps of "omni-directional" connectivity, which provide a spatial representation of connectivity between every possible pair of start and endpoints in the landscape. These maps can be used by researchers and landscape managers to understand and predict how ecological processes (e.g., animal movement, disease transmission, gene flow, and fire behavior) are likely to manifest in geographic space. More information on the broader Circuitscape project, which is home to Circuitscape.jl and Omniscape.jl, can be found at [circuitscape.org](https://circuitscape.org).
+Omniscape.jl is a software package that implements the Omniscape algorithm [@mcrae2016] to compute landscape connectivity. It is written in the Julia programming language [@bezanson2017] to be fast, scalable, and easy-to-use. Circuitscape.jl [@anantharaman2020], the package on which Omniscape.jl builds and expands, abstracts landscapes as two-dimensional electrical networks and solves for current flow. The current flow that results represents landscape connectivity. Omniscape.jl is novel in that it produces maps of "omni-directional" connectivity, which provide a spatial representation of connectivity between every possible pair of start and endpoints in the landscape. These maps can be used by researchers and landscape managers to understand and predict how ecological processes (e.g., animal movement, disease transmission, gene flow, and fire behavior) are likely to manifest in geographic space. Omniscape.jl makes use of Julia's native multi-threading, making it readily scalable and deployable to high performance compute nodes. More information on the broader Circuitscape project, which is home to Circuitscape.jl and Omniscape.jl, can be found at [circuitscape.org](https://circuitscape.org).
 
 # Motivation
 
@@ -64,7 +64,7 @@ In addition to cumulative current, Omniscape.jl also optionally provides two add
 
 # Usage
 
-Omniscape.jl is run from the Julia REPL. It offers a single user-facing function, `run_omniscape`, which has two methods. The first method accepts a single argument specifying the path to an [INI file](https://en.wikipedia.org/wiki/INI_file) that contains input data file paths and run options. The second method of `run_omniscape` accepts arrays representing resistance and other spatial data inputs, and a dictionary of arguments specifying algorithm options. A complete user guide for Omniscape.jl, including installation instructions, function documentation, examples, and a complete list of options and their defaults, can be found in the [Omniscape.jl documentation](https://docs.circuitscape.org/Omniscape.jl/latest/).
+Omniscape.jl is run from the Julia REPL. It offers a single user-facing function, `run_omniscape`, which has two methods. The first method accepts a single argument specifying the path to an [INI file](https://en.wikipedia.org/wiki/INI_file) that contains input data file paths and run options. Spatial data inputs can be in either ASCII or GeoTIFF raster formats, and outputs can also be written in either format. The second method of `run_omniscape` accepts arrays representing resistance and other spatial data inputs, and a dictionary of arguments specifying algorithm options. A complete user guide for Omniscape.jl, including installation instructions, function documentation, examples, and a complete list of options and their defaults, can be found in the [Omniscape.jl documentation](https://docs.circuitscape.org/Omniscape.jl/latest/).
 
 
 # Acknowledgments
