### 2021-01-15 9e76e62
fix figure reference
