### 2021-01-15 #2829 @vlandau
@tpoisot just added a figure with example outputs. Good call! I also added a sentence about multithreading to the summary, and mentioned that ASCII and GeoTIFF raster formats are supported for reading and writing.

@melissawm I think it might make sense prior to finalizing the review to tag a new patch release of Omniscape.jl so a snapshot of the documentation can be preserved. Would it be possible to update the version after I tag that release?
