### 2021-01-15 d2e1d1f
enable github actions on debug branch
