+++ .github/workflows/ci.yml
@@ -7,6 +7,7 @@ on:
       - main
   push:
     branches:
+      - vl/debug-#88
       - main
     tags: '*'
 jobs:
