+++ .github/workflows/ci.yml
@@ -44,8 +44,6 @@ jobs:
       - uses: julia-actions/julia-buildpkg@v1
       - name: Setup tmate session
         uses: mxschmitt/action-tmate@v3
-        with:
-          sudo: false
       - uses: julia-actions/julia-runtest@v1
       - uses: julia-actions/julia-processcoverage@v1
       - uses: codecov/codecov-action@v1
