### 2021-01-20 #2829 @melissawm
Thank you both @tpoisot and @juliohm for you thorough and thoughtful reviews!

> @melissawm I think it might make sense prior to finalizing the review to tag a new patch release of Omniscape.jl so a snapshot of the documentation can be preserved. Would it be possible to update the version after I tag that release?

If I understand the question correctly, yes you can update the version number with whedon and choose the correct version to tag as the Zenodo/archive corresponding to the JOSS paper. Please see if [the JOSS/whedon documentation](https://joss.readthedocs.io/en/latest/whedon.html#changing-the-software-version) answers your question.
