### 2021-01-20 #2829 @vlandau
@melissawm looks like I don't have the permissions to update. But v0.4.4 is tagged and docs for it are [here](https://docs.circuitscape.org/Omniscape.jl/v0.4.4), so it's all set.
