### 2021-01-20 #2829 @melissawm
@vlandau, A few final comments:
- In the reference Anantharaman et al, 2020, the publication venue listed in the paper is "Proceedings of the JuliaCon Conferences", which is different from the DOI (JuliaCon Proceedings). Can you check which is the correct one?
- For the reference Bunn et al, 2000, can you confirm that [this is the right DOI](https://doi.org/10.1006/jema.2000.0373)? In that case, can you add it to the reference?
- For the McRae et al, 2016 reference, it seems to me that [this is the right DOI](https://www.researchgate.net/publication/304842896_Conserving_Nature%27s_Stage_Mapping_Omnidirectional_Connectivity_for_Resilient_Terrestrial_Landscapes_in_the_Pacific_Northwest?channel=doi&linkId=577c4e4608aec3b743367195&showFulltext=true) - can you confirm and add it to the reference if that's the case?

After this, I ask that you re-read the paper and make sure all is good so we can proceed with publishing. Thanks!
