+++ paper/paper.bib
@@ -18,7 +18,8 @@
   number={4},
   pages={265--278},
   year={2000},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1006/jema.2000.0373}
 }
 
 @article{bezanson2017,
@@ -84,5 +85,6 @@
   title={Conserving nature’s stage: mapping omnidirectional connectivity for resilient terrestrial landscapes in the pacific northwest},
   author={McRae, Brad H and Popper, K and Jones, A and Schindel, M and Buttrick, S and Hall, K R and Unnasch, R S and Platt, J},
   journal={The Nature Conservancy, Portland, Oregon},
-  year={2016}
+  year={2016},
+  doi={10.13140/RG.2.1.4158.6166}
 }
