### 2021-01-20 #2829 @vlandau
@melissawm I've added those DOIs, thanks! I have reviewed the paper and all looks good to me!

For Anantharaman et al, I am checking with the folks from Julia. It looks like the citation shown on that page has "JuliaCon Proceedings" but the bibtex snippet that they provide on the same page uses "Proceedings of the JuliaCon Conferences." I will post back here once I find out which to use.
