+++ paper/paper.bib
@@ -78,7 +78,7 @@
   pages = {58},
   author = {Ranjan Anantharaman and Kimberly Hall and Viral B. Shah and Alan Edelman},
   title = {{Circuitscape in Julia: High performance connectivity modelling to support conservation decisions}},
-  journal = {Proceedings of the JuliaCon Conferences}
+  journal = {JuliaCon Proceedings}
 }
 
 @article{mcrae2016,
