### 2021-01-21 #2829 @vlandau
@melissawm Okay, they said to go with JuliaCon Proceedings. Thanks for flagging that! I just fixed the reference in the paper.bib file.
