### 2021-01-25 #2829 @melissawm
At this point could you:
- [ ] Archive the reviewed software in Zenodo or a similar service (e.g., figshare, an institutional repository) with the correct version

I can then move forward with accepting the submission. Thanks!
