### 2021-01-29 #2829 @melissawm
Thank you, @vlandau ! Just checking - there is a small typo on the summary for Zenodo:

> Omniscape.jl implements **to** Omniscape omnidirectional habitat connectivity modeling algorithm (McRae et al. 2016) in the Julia programming language.

Also you mention References but there is only one paper listed - is this intentional?
