### 2021-01-29 #2829 @melissawm
Hello @openjournals/joss-eics , I believe this is ready for publishing.

Thanks again @vlandau, @juliohm and @tpoisot !
